// src/types/commonEnums.ts
export enum Languages {
  BG = "BG",
  CS = "CS",
  DA = "DA",
  DE = "DE",
  EL = "EL",
  EN = "EN",
  "EN-GB" = "EN-GB",
  "EN-US" = "EN-US",
  ES = "ES",
  ET = "ET",
  FI = "FI",
  FR = "FR",
  HU = "HU",
  ID = "ID",
  IT = "IT",
  JA = "JA",
  KO = "KO",
  LT = "LT",
  LV = "LV",
  NB = "NB",
  NL = "NL",
  PL = "PL",
  PT = "PT",
  "PT-BR" = "PT-BR",
  "PT-PT" = "PT-PT",
  RO = "RO",
  RU = "RU",
  SK = "SK",
  SL = "SL",
  SV = "SV",
  TR = "TR",
  UK = "UK",
  ZH = "ZH",
}